common: force float and double arguments to logf() and log(), respectively
Fixes warnings on some MSVC versions

+++ source/common/common.h
@@ -102,8 +102,8 @@
 #endif // if ENABLE_CYCLE_COUNTERS
 
 #if defined(_MSC_VER)
-#define X265_LOG2F(x) (logf(x) * 1.44269504088896405f)
-#define X265_LOG2(x) (log(x) * 1.4426950408889640513713538072172)
+#define X265_LOG2F(x) (logf((float)(x)) * 1.44269504088896405f)
+#define X265_LOG2(x) (log((double)(x)) * 1.4426950408889640513713538072172)
 #else
 #define X265_LOG2F(x) log2f(x)
 #define X265_LOG2(x)  log2(x)
